% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ad.gent}
\alias{ad.gent}
\title{Gene-based results for Alzheimer's disease}
\format{
This dataset contains results from genome-wide gene-based association testing applied to Alzheimer's disease
\describe{
\item{gene}{HGNC gene symbol}
\item{chr}{Chromosome}
\item{mid}{Midpoint of gene body in hg19 coordinates}
\item{pval}{Gene-based association test P-value from GenT}
\item{m}{Number of SNPs used in gene-based association testing}
\item{gent_sigma2_h0}{Variance of the gene-based association test statistic under the null hypothesis, which is 2 times the trace of the squared LD matrix of SNPs used}
\item{gent_test_statistic}{Gent-based association test statistic from GenT, i.e. the sum of SNP chi-square statistics}
}
}
\source{
The Alzheimer's disease GWAS was from Bellenguez et al., 2021 of 480K subjects and the LD reference panel was the 1000 Genomes Phase 3 European panel
}
\usage{
ad.gent
}
\description{
Gene-based association test results for Alzheimer's disease
}
\examples{
data(ad.gent)
head(ad.gent)
}
\keyword{datasets}
