% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gent.Rho}
\alias{gent.Rho}
\title{Gene-based test statistic correlations}
\format{
This dataset is a list where each element is a matrix of correlations between gene-based test statistics due to LD between SNPs. Rows/columns of these matrices are named by the HGNC symbol of the gene to which the row/column corresponds. Rows/columns are ordered by base pair location of the transcription start site of each gene in hg19 coordinates.
}
\source{
Genes tested using the 1000 Genomes Phase 3 European reference panel
}
\usage{
gent.Rho
}
\description{
Each entry in this list is a matrix of correlations between gene-based test statistics.
}
\examples{
data(gent.Rho)
str(my_data)
}
\keyword{datasets}
