% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbd.gent}
\alias{lbd.gent}
\title{Gene-based results for Lewy body dementia}
\format{
This dataset contains results from genome-wide gene-based association testing applied to Lewy body dementia
\describe{
\item{gene}{HGNC gene symbol}
\item{chr}{Chromosome}
\item{mid}{Midpoint of gene body in hg19 coordinates}
\item{pval}{Gene-based association test P-value from GenT}
\item{m}{Number of SNPs used in gene-based association testing}
\item{gent_sigma2_h0}{Variance of the gene-based association test statistic under the null hypothesis, which is 2 times the trace of the squared LD matrix of SNPs used}
\item{gent_test_statistic}{Gent-based association test statistic from GenT, i.e. the sum of SNP chi-square statistics}
}
}
\source{
The Lewy body dementia GWAS was from Chia et al., 2021 of 16516 subjects and the LD reference panel was the 1000 Genomes Phase 3 European panel
}
\usage{
lbd.gent
}
\description{
Gene-based association test results for Lewy body dementia
}
\examples{
data(lbd.gent)
head(lbd.gent)
}
\keyword{datasets}
