% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ld.df}
\alias{ld.df}
\title{Gene LD scores}
\format{
This dataset contains allele frequency-weighted sums of gene-specific LD scores, and their squared versions, for 17479 genes spanning th 22 autosomes.
\describe{
\item{chr}{Chromosome}
\item{gene}{HGNC gene symbol}
\item{m}{Number of SNPs whose LD scores were summed for the gene}
\item{sumwldscore}{Sum of MAF-weighted LD scores for the gene}
\item{sumwld2score}{Sum of MAF-weighted squared LD scores for the gene}
}
}
\source{
LD scores were calculated from the 1000 Genomes Phase 3 European reference panel in +-1Mb windows around gene bodies in hg19 coordinates
}
\usage{
ld.df
}
\description{
MAF-weighted gene-specific LD scores
}
\examples{
data(ld.df)
head(ld.df)
}
\keyword{datasets}
