% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{posterior_gene}
\alias{posterior_gene}
\title{This function returns gene-level posterior probabilities of causality for a single trait that do not consider correlation between genes}
\usage{
posterior_gene(
  gent.data,
  gwasn,
  prior.estimation.list,
  ld.df = data(ld.df),
  gene_id = "gene",
  chromosome = "chr",
  position = "mid",
  pval = "pval",
  null_mean = "m",
  null_variance = "gent_sigma2_h0",
  gamma_shape = "shape",
  gamma_rate = "rate",
  verbose = TRUE
)
}
\arguments{
\item{gent.data}{data frame of full gene-based association test results}

\item{gwasn}{GWAS sample size}

\item{prior.estimation.list}{direct output of the compositemh function containing information regarding the prior distributions to use in posterior estimation}

\item{ld.df}{data frame of weighted LD scores for all genes in gent.data}

\item{verbose}{TRUE if progress should be printed, FALSE otherwise}
}
\description{
This function returns gene-level posterior probabilities of causality for a single trait that do not consider correlation between genes
}
\examples{
posterior_gene()
}
\keyword{posterior}
